+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Additional weather infos (pressure, humidity and wind)
+- Option to display "feels like" temperature instead
 
 ### Changed